bugfix: shim_do_pool incorrectly setting revents wrt POLLERR and POLLHUP
This is issue #389 on the Graphene project.

The function shim_do_poll has an error on the line that reads:

    fds[i].revents |= (fds[i].events & (POLLERR|POLLHUP));

Per the manpage, POLLERR and POLLUP are returned in `revents` and
ingnored in `events`.  Indeed, programs hsould not be setting, and
typically don't set -- `events` with these values.

The effect of this bug is, among others, that a process will never get a
disconnect evetn fro a peer socket.