bugfix: rename could corrupt handles that are open to old file
This is Issue #382 on the grapheen project.

If an application holds a handle for an open file foo, and foo is
renatemd to bar, then the handle's data is modfiefy by chroot_rename.
In particular, the statement in LibOS/shim/src/fs/chroot/fs.c's function
chroot_rename:

    atomic_set(&old->data->size, 0)

means that the old handle is not to a zero-sized file.

The effects of this bug are particularly pernicious if the handle is
sued as the input hadnle arguemtn to senfile.  In particular, in
sys/shim_fs.c, shim_do_senfile calls handle_copy.  In handle_copy,
chorot fs's poll callabkc is called to get he size of the file
(which has been erroneoulsy set to 0), causing the count vairble to
be set to `size - offi` -- a negative number.

I don't fully understand the reference counting and cacheing of
handles that graphene does, so consider this fix to be a "dirty fix":
it fixes functionality, but might be leaking memory.