bugfix: epoll, poll, and select do not work with newly created sockets
This is a quick and dirty bugfix for graphene issue #365.

Due to the minimal PAL interface, which foces socket()+connect() and
socket()+bin()+listen() into a single function (DkStreamOpen), Graphene
cannot create a host socket descritpor at the application's call to
socket(), but must rather wait until the application calls connect() or
bind().

Some applications, such as nginx, implement a non-blocking connect by
first calling socket(), then addding that descritpor to the
epoll/poll/select list, and then calling connect() when these
functions indiciate that the descriptor is readable/writeable (which is
trivialy true, at least initially for the fd).

The problem is that, when adding such an fd via say,
epoll_ctl(EPOLL_CTL_ADD), epoll_ctl will return EPREM, since
pal_handle has not yet been associated with graphene's internal hdl for
that decriptor.

This bugfix only addresse epoll.  The fix's approach is that the
implemenation for epoll_ctl checks if the application is adding an fd
with a NULL pal_hdl (which is the case in the connect scenario we
described above).  If so, the pal_hdl is set to a hardcoded
magic value.  The implementation for epoll_wait first checks the list of
fds to see if any have this magic value; if so, epoll_wait() returns
that those fd's are readable and writeable.