Change CA config; nextfs and mdish support moutable mounts.
The previous approach of having the configuraiotn entry
pheonix.cafile point to file on a chroot mount is problematic
for two reasons: (1) the chroot mount must be mounted first, and
(2) the data structure storing the config is not yet
reconstituted on the fs's migrate callback.

To addresss these issues, nextfs and mdish now look for a config
entry called pheonix.ca_der, which is the hex encoding of the CA
DER certificate.  The fs's mount callback can read this entry
without problem and store the DER as private mount data.
Forked children must retrieve this value from the mount data that
is copied over during a fork/migration; note that the mount
callback is invoked for the child.

To allow nextfs and mdish to support multiple mounts, I moved
the global nextfs_client and mdish_client into the private mount
data as well.  This presents a small challenge, as each fs
function must now retrieve the client by following the pointers
in one of the function's arguments, which is typically either a
shim_handle or shim_dentry until the private mount data is
retrieved.