smdish: use a global for the mount data
I have been retrieving the mount data by following
pointers in the file handle; I was concerned that,
during fork, the file handle's pointer to the mount
data was not being updated for the child.  I don't think
this was actually happening, however.

In any case, using the global is cleaner.