smdish fs: fix migration of private segment replica
The issue is that bookkeep/shim_vma.c's BEGIN_CP_FUNC(vma)
function will only migrate the private (in-enclave) segment
replica if it thinks the size of the file is greater than 0.
It tests this by calling get_file_size, which calls into the
fileystems's hstat() function handler.  Previously, we did not
implement hstat() for smdish.  Thus, the memory replica does not
get migrated over.  Moreover, to make things more confusing,
since the memory is not migrated, on restore in the child, the
BEGIN_RS_FUNC(vma) would call smdish_mmap().

We now implement hstat() to return the stat.st_size as the
size of the mapping; this causes the memory to migrate over and
prevents smdish_mmap() from being recalled for each segment on
restore.

Note that, because NGINX stores the lock information (such as
the lock's name) in the shared memory segment, a child process
always needs to have a version of the replica so that it can
safely call lock() and get the updated version of the replica;
otherwise, it won't even know the name of the path to pass to
lock().