mdish fs: fix migration; fix encrypting data in place.
The migration issue is the same as the previous commit for smdish.

The encryption issue is due to the fact that BearSSL does
encryption on a buffer in place.  This would mean that we would
encrypt the private replica and copy it to the public replica
on unlock().  Unfortunately, NGINX stores the lock name in the
shared segment, so encryption in place means the process would not
know the plaintext lock name on the next call to lock().  Thus, we know
keep the private segment in plaintext, at the loss of an extra memcpy.