rename: band-aid fix to handle common case.
Graphene's rename() syscall is broken.  The issue that this
commit fixes is that the new files mode and type were
not being copied from the old files, and were left as unitialized
values.  Thus, cases the filesystem-specific rename handler to
not be called, and resulted in a new file that was of 0-length and
had a bogus type and mode.  This comment copies the mode and type
from the old file.