smdish: remove check for re-mmap'ing.
Before the smdish_hstat() function was implementated, mmap()
would be recalled on migration.  I added a check to see if this
was the case.  Note that this didn't solve the issue with the
private segment replica data not being migrated; that was only
solved after I implemented smdish_hstat().

In other words, while tracking down the migration issues, I
added some code in smdish_mmap() that, in the end, didn't serve
a purpose.