Initial merge/update of Phoenix with Graphene master (3ef4e96)
This is the initial commit which attempts to merge the Phoenix changes
with the current Graphene master.  Subsequent commits will clean up the
code further.

Phoenix was a hard fork of Graphene at the time when Graphene
was very modestly maintained.  As Graphene became more actively
maintained, I tried to backport the Graphene master changes
into Phoenix, but it became untenable.  One of the biggest issues
was that I had applied an exitless patch, which changed quite
a few files in Pal/; Graphene had not applied this patch, and moreover,
was actively refactoring parts of Pal.

Graphene eventually merged in the exitless feature.  Now, it just
seems easier to branch off of Graphene master, apply my changes, and
then periodically merge in any updates from Graphene master.

My changes are fairly isolated to:

- rho and rpc utility functions:
  - LibOS/shim/include/rho_*, and LibOS/shim/include/rpc.h
  - LibOS/shim/src/utils/rho_*, and LibOS/shim/src/utils/rpc.c

- new filesystems:
  - LibOS/shim/src/fs/nextfs
  - LibOS/shim/src/fs/smc
  - LibOS/shim/src/fs/smdish
  - LibOS/shim/src/fs/smuf

- modifications to the fcntl and time system calls:
  - LibOS/shim/src/sys/shim_fcntl.c
  - LibOS/shim/src/sys/shim_time.c

- The addition of BearSSL for SSL and crypto functionality:
  - bearssl-0.6/

- The addition of a few libc functions to Pal/lib.

As well as minor changes to the associated Makefiles and headers.